Fixes possible ioverflow by expanding the range returned.
Since some functions need to return a negative value, and Windows
uses 'ints' (which may or may not be 64 bits) for return values
of its functions, I made it all i64 all around.

The parameters are still u32, though, so the callers will not be
able to send more than a u32 in a single call.